+++ libs/elements/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.3.0 (2020-01-14)
+
+### Bug Fixes
+
+- **elements:** 修复 code-diff-editor value 未及时更新错误 ([23b1177](https://github.com/aiao-io/aiao/commit/23b1177def7304bef5b71457235388620b6d9007))
+- **elements:** 修复 code-editor value 未及时更新错误 ([95f201c](https://github.com/aiao-io/aiao/commit/95f201c8c022d980ba9168db5165ca99c017e15b))
+
+# 0.2.0 (2020-01-14)
+
+### Features
+
+- **elements:** add code-diff-editor ([#56](https://github.com/aiao-io/aiao/issues/56)) ([b8eae66](https://github.com/aiao-io/aiao/commit/b8eae6661d2784fe77f4106cbb1aedab02bfc759))
+- **elements:** code-editor support json object ([#55](https://github.com/aiao-io/aiao/issues/55)) ([4a9eed7](https://github.com/aiao-io/aiao/commit/4a9eed7dba3478805674c35254e24be55acb6663))
+
+### Performance Improvements
+
+- **elements:** 改进了 img 组件加载图片速度 ([#53](https://github.com/aiao-io/aiao/issues/53)) ([cc57233](https://github.com/aiao-io/aiao/commit/cc572332fe5df382b45166e3148e0e860c60b802))
+
 # 0.1.0 (2020-01-01)
 
 ### Features